% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scicosa.R
\name{PRIMITIVE.trigrid}
\alias{PRIMITIVE.trigrid}
\title{Primitive from icosa}
\usage{
PRIMITIVE.trigrid(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Primitive from icosa
}
\examples{
#library(tibble)
#library(icosa)
#tri <- trigrid()
#library(sc)
#library(scgraph)
#scgraph::as.igraph(PRIMITIVE(tri))
}
